requirements.txt
# Base system
Django==1.11.4
psycopg2==2.7.3
Unipath==1.1
dj-database-url==0.4.2
djangorestframework==3.6.3
django-dotenv==1.4.1
Pillow==4.2.1

# Useful libs
django-messagegroups==0.4.5
django-model-utils==3.0.0
django-bootstrap-form==3.3
sorl_thumbnail==12.3

# Dev
django-extensions==1.8.1

# Testing
pytest==3.2.1
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-pep8==1.0.6
model-mommy==1.3.2

# Deployment
gunicorn==19.7.1
opbeat==3.5.2
